Altera package das entidades
